project/webClientProject/requirements.txt
Package       Version
------------- -------
asgiref       3.3.1  
Django        3.1.7  
paho-mqtt     1.5.1  
pip           20.0.2 
pkg-resources 0.0.0  
pytz          2021.1 
setuptools    44.0.0 
sqlparse      0.4.1  
